Move MainWindow out of ui.py and rename it to TabbedWindow
